pyproject.toml
[project]
name = "books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "drf-spectacular[sidecar]>=0.27.2",
    "psycopg>=3.2.3",
    "whitenoise>=6.7.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "faker>=30.8.1",
]
